# Alternative fix for #62 with speedup