@elithrar 2016-01-24
This looks good from a high-level POV. I haven't had a chance to dive in too deeply, but are there any cases where the new regex would break existing behavior?

(not wrapping existing groups is a nice touch)