@elithrar 2016-02-28
Alright:

* This looks solid—there's a good performance boost here (although router performance isn't a bottleneck for 99% of applications)
* Doesn't break existing tests
* No objections in a month.

Going to merge. Thanks for the work on this @tumdum!