@jprobinson 2016-03-07
Hi guys. I haven't figured out which change did it exactly, but this commit broke one of my project's builds today.

The test that broke has this as the route:

`/v2/puzzle/{type:(?i:daily|mini|.*?bonus|.*?variety)}-{date:\\d{4,4}-\\d{2,2}-\\d{2,2}}.json`

It _still_ matches this route:

`/v2/puzzle/daily-2015-02-01.json`

But the value from "date" is getting put into the "type" value in the map returned by `mux.Vars`.

If it helps, this path is registered to a `mux.PathPrefix("/svc/xxxx").Subrouter()`.