@elithrar 2016-03-07
Resolved in https://github.com/gorilla/mux/pull/155